third_party/blink/web_tests/platform/generic/fast/dom/collection-null-like-arguments-expected.txt

This test reports the value that document.all returns for various null-like arguments.

Written for http://bugs.webkit.org/show_bug.cgi?id=13106.

document.all[0]:      [object HTMLHtmlElement]
document.all(0):      [object HTMLHtmlElement]
document.all.item(0): [object HTMLHtmlElement]

document.all[""]:      undefined
document.all(""):      null
document.all.item(""): null

document.all["0"]:      [object HTMLHtmlElement]
document.all("0"):      [object HTMLHtmlElement]
document.all.item("0"): [object HTMLHtmlElement]

document.all[undefined]:      undefined
document.all(undefined):      null
document.all.item(undefined): null

document.all[null]:      undefined
document.all(null):      null
document.all.item(null): null
